Use unique contexts for installer library lookup
The use of a constant context meant that each call to
`temporaryConfiguration` resulted in the same configuration being
returned.

As `temporaryConfiguration` only adds the requested libraries on the
first use of the configuration, any subsequent invocation with the same
context, even with different dependencies as arguments, would just
return the configuration from the first invocation unchanged.